+++ noodles-fasta/examples/fasta_index.rs
@@ -11,7 +11,7 @@ use noodles_fasta::{self as fasta, fai};
 fn main() -> io::Result<()> {
     let src = env::args().nth(1).expect("missing src");
 
-    let index = fasta::io::index(src)?;
+    let index = fasta::fs::index(src)?;
 
     let stdout = io::stdout().lock();
     let mut writer = fai::Writer::new(stdout);